migration fixes

+++ db/migrate/003_extend_sites.rb
@@ -1,9 +1,13 @@
 class ExtendSites < ActiveRecord::Migration
   def self.up
-    add_column :sites, :forum_layout_id, :integer
+    if defined? Site
+      add_column :sites, :forum_layout_id, :integer
+    end
   end
 
   def self.down
-    remove_column :sites, :forum_layout_id
+    if defined? Site
+      remove_column :sites, :forum_layout_id
+    end
   end
 end
